Rename[]: 추가 view 패키지
맵액티비티와 서치프래그먼트를 뷰 패키지로 이동시킴
